#13 migration process refactoring

+++ src/db/migrations/migration-2020-05-25-11-18-00-convertData.js
@@ -20,11 +20,14 @@ const convertData = () => new Promise(resolve => {
     });
 
     rl.on('close', () => {
-      fs.unlinkSync('log.txt');
-      fs.renameSync('log1.txt', 'log.txt');
+      if (count) {
+        fs.unlinkSync('log.txt');
+        fs.renameSync('log1.txt', 'log.txt');
+      }
+
       resolve();
     });
-  }
+  } else { resolve(); }
 });
 
 module.exports = {
